refactor: 템플릿-콜백 패턴을 활용하여 리팩토링
