+++ jdbc/src/main/java/org/springframework/jdbc/core/JdbcTemplate.java
@@ -1,37 +1,23 @@
 package org.springframework.jdbc.core;
 
-import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.dao.IncorrectResultSizeDataAccessException;
 
 public class JdbcTemplate {
 
-    private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
-
-    private final DataSource dataSource;
+    private final QueryTemplate queryTemplate;
 
     public JdbcTemplate(final DataSource dataSource) {
-        this.dataSource = dataSource;
+        this.queryTemplate = new QueryTemplate(dataSource);
     }
 
     public void update(String sql, Object... args) {
-        try (
-            Connection conn = dataSource.getConnection();
-            PreparedStatement prepareStatement = setUpPreparedStatement(conn, sql, args)
-        ) {
-            prepareStatement.executeUpdate();
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
+        queryTemplate.service(sql, PreparedStatement::executeUpdate, args);
     }
 
     public <T> T queryForObject(String sql, RowMapper<T> rowMapper, Object... args) {
@@ -50,27 +36,13 @@ public class JdbcTemplate {
     }
 
     public <T> List<T> query(String sql, RowMapper<T> rowMapper, Object... args) {
-        try (
-            Connection conn = dataSource.getConnection();
-            PreparedStatement prepareStatement = setUpPreparedStatement(conn, sql, args);
+        return queryTemplate.service(sql, prepareStatement -> {
             ResultSet resultSet = prepareStatement.executeQuery();
-        ) {
             List<T> result = new ArrayList<>();
             while (resultSet.next()) {
                 result.add(rowMapper.mapRow(resultSet));
             }
             return result;
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
-    }
-
-    private PreparedStatement setUpPreparedStatement(Connection conn, String sql, Object... args) throws SQLException {
-        PreparedStatement prepareStatement = conn.prepareStatement(sql);
-        for (int i = 0; i < args.length; i++) {
-            prepareStatement.setObject(i + 1, args[i]);
-        }
-        return prepareStatement;
+        }, args);
     }
 }
