+++ jdbc/src/main/java/org/springframework/jdbc/core/QueryCallback.java
@@ -1,0 +1,10 @@
+package org.springframework.jdbc.core;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface QueryCallback<T> {
+
+    T execute(PreparedStatement preparedStatement) throws SQLException;
+}
